+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Bug Fixes
 - fix "To date" computation to include the selected day (also fix the top computation).
 - fix project page jumbotron link cut when name is multi words.
 - fix infos/contributor endpoint that ignored dto/dfrom.
+- fix groups endpoint that ignored dto/dfrom.
 
 Other Notes
 -----------